requirements.txt
python-dotenv
requests
google-cloud-vision
Pillow
tqdm